Type DataLoader
Add types to the DataLoader. The main drive was to get batch_load_fn to
allow the type engine to infer the type of load, and this accomplishes
that. Two unfortunate side-effects:

One is that I don't know how to type it to accept any Sequence type (or,
at the bare minimum, Tuple | List), because they're generally
type-invariant, so even if it's easy to see that batch_load_fn will
return a sequence type that matches what it gets, pyright still won't
allow assignment because DataLoader has to be typed with a Union to also
accept the Generic type, and there's no way that I know to separately
TypeVar a Generic type and its parameter.

Two is that I don't know how to type instance properties (or, for that
matter, Generics) with 3.5-style comment hints, so this will syntax
error in 3.5. Happy to convert it over if anyone knows how.